Revert "Rename `borrow::internal` to `borrow::private`"
This reverts commit 9e6ba88d373d3c63cfb20d8a8ec0bf16f2abda32.
