+++ src/borrow/mod.rs
@@ -31,18 +31,19 @@
 //! [ArrayBuffer]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/ArrayBuffer
 //! [Buffer]: https://nodejs.org/api/buffer.html
 
-pub(crate) mod private;
+pub(crate) mod internal;
 
 use std::fmt;
 use std::ops::{Deref, DerefMut, Drop};
 use std::os::raw::c_void;
 
+use self::internal::Pointer;
 use crate::context::Lock;
 
 /// A trait for JS values whose internal contents can be borrowed immutably by Rust while the JS engine is locked.
 pub trait Borrow: Sized {
     /// The type of the value's internal contents.
-    type Target: private::Pointer;
+    type Target: Pointer;
 
     /// Borrow the contents of this value immutably.
     ///
@@ -101,12 +102,12 @@ impl fmt::Display for LoanError {
 }
 
 /// An immutable reference to the contents of a borrowed JS value.
-pub struct Ref<'a, T: private::Pointer> {
+pub struct Ref<'a, T: Pointer> {
     pointer: T,
     lock: &'a Lock<'a>,
 }
 
-impl<'a, T: private::Pointer> Ref<'a, T> {
+impl<'a, T: Pointer> Ref<'a, T> {
     pub(crate) unsafe fn new(lock: &'a Lock<'a>, pointer: T) -> Result<Self, LoanError> {
         let mut ledger = lock.ledger.borrow_mut();
         ledger.try_borrow(pointer.as_ptr())?;
@@ -114,14 +115,14 @@ impl<'a, T: private::Pointer> Ref<'a, T> {
     }
 }
 
-impl<'a, T: private::Pointer> Drop for Ref<'a, T> {
+impl<'a, T: Pointer> Drop for Ref<'a, T> {
     fn drop(&mut self) {
         let mut ledger = self.lock.ledger.borrow_mut();
         ledger.settle(unsafe { self.pointer.as_ptr() });
     }
 }
 
-impl<'a, T: private::Pointer> Deref for Ref<'a, T> {
+impl<'a, T: Pointer> Deref for Ref<'a, T> {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {
@@ -130,12 +131,12 @@ impl<'a, T: private::Pointer> Deref for Ref<'a, T> {
 }
 
 /// A mutable reference to the contents of a borrowed JS value.
-pub struct RefMut<'a, T: private::Pointer> {
+pub struct RefMut<'a, T: Pointer> {
     pointer: T,
     lock: &'a Lock<'a>,
 }
 
-impl<'a, T: private::Pointer> RefMut<'a, T> {
+impl<'a, T: Pointer> RefMut<'a, T> {
     pub(crate) unsafe fn new(lock: &'a Lock<'a>, mut pointer: T) -> Result<Self, LoanError> {
         let mut ledger = lock.ledger.borrow_mut();
         ledger.try_borrow_mut(pointer.as_mut())?;
@@ -143,14 +144,14 @@ impl<'a, T: private::Pointer> RefMut<'a, T> {
     }
 }
 
-impl<'a, T: private::Pointer> Drop for RefMut<'a, T> {
+impl<'a, T: Pointer> Drop for RefMut<'a, T> {
     fn drop(&mut self) {
         let mut ledger = self.lock.ledger.borrow_mut();
         ledger.settle_mut(unsafe { self.pointer.as_mut() });
     }
 }
 
-impl<'a, T: private::Pointer> Deref for RefMut<'a, T> {
+impl<'a, T: Pointer> Deref for RefMut<'a, T> {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {
@@ -158,7 +159,7 @@ impl<'a, T: private::Pointer> Deref for RefMut<'a, T> {
     }
 }
 
-impl<'a, T: private::Pointer> DerefMut for RefMut<'a, T> {
+impl<'a, T: Pointer> DerefMut for RefMut<'a, T> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.pointer
     }
