+++ src/macro_internal/mod.rs
@@ -1,5 +1,5 @@
 //! Internals needed by macros. These have to be exported for the macros to work
-pub use crate::context::private::{initialize_module, Env};
+pub use crate::context::internal::{initialize_module, Env};
 /// but are subject to change and should never be explicitly used.
 
 #[cfg(feature = "legacy-runtime")]
