+++ src/types/binary.rs
@@ -1,8 +1,8 @@
 //! Types and traits representing binary JavaScript data.
 
-use crate::borrow::private::Pointer;
+use crate::borrow::internal::Pointer;
 use crate::borrow::{Borrow, BorrowMut, LoanError, Ref, RefMut};
-use crate::context::private::Env;
+use crate::context::internal::Env;
 use crate::context::{Context, Lock};
 #[cfg(feature = "napi-1")]
 use crate::handle::Handle;
